Add files via upload
I created this project to use a Random Forest Classification model that can be used to predict Customer Churn  in a company. This Model can thus help companies to solve their churn problem as it can make predictions that are 92% accurate in the Base model. 
The complete code is attached in the ipynb file. 
**No EDA was done in this analysis